구조개편 및 트라이포드 인식부 추가 1. 화면캡쳐와 이미지 분석 클래스 구조개편  - 트라이포드가 추가되면서 기존 화면캡쳐 호출 방식은 스레드가 하나 추가로 생기며 리소스 사용량이 많아져 해당 구조를 개편함 2. 트라이포드 인식부 추가  - LL프로젝트 기능사항 두번째인 트라이포드파트 추가
