+++ LostarkLogProject/Properties/Settings.settings
@@ -11,5 +11,8 @@
     <Setting Name="트레이로실행" Type="System.Boolean" Scope="User">
       <Value Profile="(Default)">False</Value>
     </Setting>
+    <Setting Name="UID" Type="System.String" Scope="User">
+      <Value Profile="(Default)" />
+    </Setting>
   </Settings>
 </SettingsFile>
